A refactor.
A performance pivot. I've gone and made it so that finalizers run at the
end of every loop interation in Cadence, but the first few steps to so
this in Hotspot shows how much performance is effected by just a single
branch. Assignments do not seem to be as bad performance wise, so I'm
considering how to make it so that looping is single assignment, instead
of a lot of branching, let's move ourselves to the head of the cadence
where there is a special step that runs finalizers if any exist, and it
will exit by simply saying that it is waiting if there is nothing more
to do. We're riding on conditions that already exisit in the loop. Maybe
that makes them hotter somehow. We shall see. I get a sense that this is
going to reduce the number of checks, but I'm not certain.

Because we are jumping a lot, we can put some of the logic inside the
steps themselves, repeating it as necessary, so that variable processing
could be performed inside a wrapper. That variable processing could be
duplicated in each function wrapper, which means it will get caught by
minzip, if we're lucky, or it can be included and hopefully inlined.

That way, we don't pay for the test and the looping if we don't use it,
which probably doens't matter, because the only time we don't use it is
when something bad happens.

Offloading thoughts so I can go work on something else.